\name{girafe-internal}
\alias{alongChromTicks}
\alias{featureColors}
\alias{plotAlongChromLegend}
\alias{newVP}
\alias{writeExportData}
\alias{getReadPosDf}
\alias{getChromLengths}
\alias{coverageOneStrand}
\title{Internal girafe functions}
\description{
  Called internally by other \code{girafe} functions.
  Normally need not be called by the user.
}
\author{Wolfgang Huber, Joern Toedling}
\seealso{\code{\link{AlignedGenomeIntervals-class}}}
\keyword{internal}
